+++ src/keys.tcl
@@ -211,22 +211,22 @@ proc keys::ReservedAdd {} {
   namespace upvar ::alited al al
   Add action exit-app     Alt-F4 {alited::Exit; break}
   Add action find-replace Control-F {alited::find::_run; break}
-  Add action find-unit    Shift-Control-F {::alited::find::FindUnit; break}
-  Add action new-file     Control-N {::alited::file::NewFile; break}
-  Add action open-file    Control-O {::alited::file::OpenFile; break}
-  Add action save-all     Shift-Control-S {::alited::file::SaveAll; break}
-  Add action save-close   Control-W {::alited::file::SaveAndClose; break}
-  Add action close-delete Control-Alt-W {::alited::file::CloseAndDelete; break}
+  Add action find-unit    Shift-Control-F {alited::find::FindUnit; break}
+  Add action new-file     Control-N {alited::file::NewFile; break}
+  Add action open-file    Control-O {alited::file::OpenFile; break}
+  Add action save-all     Shift-Control-S {alited::file::SaveAll; break}
+  Add action save-close   Control-W {alited::file::SaveAndClose; break}
+  Add action close-delete Control-Alt-W {alited::file::CloseAndDelete; break}
   Add action help         F1 {alited::tool::Help}
   # other keys are customized in Preferences
   Add action save-file    [alited::pref::BindKey 0 - Control-S] ::alited::file::SaveFile
-  Add action save-as      [alited::pref::BindKey 1 - Alt-S] {::alited::file::SaveFileAs; break}
+  Add action save-as      [alited::pref::BindKey 1 - Alt-S] {alited::file::SaveFileAs; break}
   Add action e_menu       [alited::pref::BindKey 2 - F4] alited::tool::e_menu3
   Add action run          [alited::pref::BindKey 3 - F5] alited::tool::_run
-  Add action indent       [alited::pref::BindKey 6 - Control-I] {::alited::edit::Indent; break}
-  Add action unindent     [alited::pref::BindKey 7 - Control-U] {::alited::edit::UnIndent; break}
-  Add action comment      [alited::pref::BindKey 8 - Control-bracketleft] {::alited::edit::Comment; break}
-  Add action uncomment    [alited::pref::BindKey 9 - Control-bracketright] {::alited::edit::UnComment; break}
+  Add action indent       [alited::pref::BindKey 6 - Control-I] {alited::edit::Indent; break}
+  Add action unindent     [alited::pref::BindKey 7 - Control-U] {alited::edit::UnIndent; break}
+  Add action comment      [alited::pref::BindKey 8 - Control-bracketleft] {alited::edit::Comment; break}
+  Add action uncomment    [alited::pref::BindKey 9 - Control-bracketright] {alited::edit::UnComment; break}
   Add action find-next    [alited::pref::BindKey 12 - F3] alited::find::FindNext
   Add action look-declaration    [alited::pref::BindKey 13 - Control-L] "::alited::find::LookDecl ; break"
   Add action look-word    [alited::pref::BindKey 14 - Control-Shift-L] "::alited::find::SearchWordInSession ; break"